fixed namings, type, comments, architecture
